+++ modules/reminders/send.ts
@@ -23,7 +23,7 @@ import {
 	remindersDatabase,
 	type Reminder,
 } from "./misc.js";
-import sendQOTD from "./qotd.js";
+import sendQuestion from "../polls/qotd.js";
 
 let nextReminder: NodeJS.Timeout | undefined;
 export default async function queueReminders(): Promise<NodeJS.Timeout | undefined> {
@@ -62,8 +62,6 @@ async function sendReminders(): Promise<NodeJS.Timeout | undefined> {
 				case SpecialReminders.Weekly: {
 					if (!channel?.isTextBased()) continue;
 
-					const date = new Date();
-					date.setUTCDate(date.getUTCDate() - 7);
 					const nextWeeklyDate = new Date(reminder.date);
 					nextWeeklyDate.setUTCDate(nextWeeklyDate.getUTCDate() + 7);
 
@@ -207,7 +205,18 @@ async function sendReminders(): Promise<NodeJS.Timeout | undefined> {
 					continue;
 				}
 				case SpecialReminders.QOTD: {
-					await sendQOTD();
+					if (!channel?.isThreadOnly()) continue;
+					remindersDatabase.data = [
+						...remindersDatabase.data,
+						{
+							channel: channel.id,
+							date: reminder.date + 86_400_000,
+							reminder: undefined,
+							id: SpecialReminders.QOTD,
+							user: client.user.id,
+						},
+					];
+					await sendQuestion(channel);
 					continue;
 				}
 			}
